fix: item image 규격 맞지 않는 부분 수정 (#106)
* fix: item image 규격 맞지 않는 부분 수정

* chore: 잘못 삭제한 css 속성 재작성

* fix: grid, flex 다른 방법으로 해결
