+++ src/mocks/data/products.json
@@ -3,7 +3,7 @@
   "products": [
     {
       "id": 1,
-      "name": "꼬북칩",
+      "name": "꼬북칩 꼬북칩 꼬북칩 꼬북칩 꼬북칩 꼬북칩 꼬북칩 ",
       "price": 1500,
       "image": "https://arqachylpmku8348141.cdn.ntruss.com/app/product/mst_product/8801056232979_L.jpg",
       "averageRating": 4.5,
